React upgrade (#829)
* Added modal

* AltinnButton added

* Updated with post create service and get organizations

* update

* Fix popper for validation

* Added regex check

* rename file

* added functionality to suggest reponame

* Service named saved to textfile

* Added spinner

* codesmell fix

* merge master

* trying to fix failing unit tests

* fix jest tests

* Pull request fixes

* reorganized folders

* added scroll to modal

* upgraded react

* ran lerna hoist

* updated snapshot

* Added altinnSpinner

* fixed jest tests
