+++ src/AltinnCore/Common/Languages/ini/en.ini
@@ -45,6 +45,7 @@ field_cannot_be_empty = Field cannot be empty
 service_name_has_illegal_characters = Service name is illegal. Service name can only contain alphanumeric and underscore ('_') characters, and must begin with a alphanumeric character
 service_name_is_too_long = Service name must contain at most 100 characters
 service_name_already_exist = Service name is already in use
+creating_your_service = Your service is being created
 
 [sync_header]
 fetch_changes = Fetch changes
