+++ src/AltinnCore/Common/Languages/ini/nb.ini
@@ -45,6 +45,7 @@ field_cannot_be_empty = Dette feltet kan ikke være tomt
 service_name_has_illegal_characters = Lagringsnavnet er ugyldig. Lagringsnavnet kan kun inneholde alfanumeriske og understrek ('_') tegn, og det må begynne på en bokstav.
 service_name_is_too_long = Lagringsnavnet kan maks bestå av 100 tegn
 service_name_already_exist = Lagringsnavnet finnes allerede
+creating_your_service = Oppretter tjenesten din
 
 [sync_header]
 fetch_changes = Hent endringer
