+++ src/react-apps/applications/dashboard/package.json
@@ -22,8 +22,8 @@
     "babel-core": "6.26.3",
     "immutability-helper": "2.7.0",
     "moment": "^2.22.2",
-    "react": "16.6.3",
-    "react-dom": "16.6.3",
+    "react": "16.7.0",
+    "react-dom": "16.7.0",
     "react-grid-layout": "0.16.6",
     "react-hot-loader": "4.3.1",
     "react-modal": "3.4.5",
@@ -41,8 +41,8 @@
   "devDependencies": {
     "@types/enzyme": "3.1.15",
     "@types/jest": "^23.3.9",
-    "@types/react": "16.4.14",
-    "@types/react-dom": "16.0.7",
+    "@types/react": "16.7.18",
+    "@types/react-dom": "16.0.11",
     "@types/react-modal": "3.1.2",
     "@types/react-redux": "6.0.2",
     "@types/react-router-dom": "4.3.1",
@@ -70,6 +70,7 @@
     "prettier": "1.13.5",
     "prettier-tslint": "^0.4.0",
     "react-test-renderer": "^16.4.0",
+    "react-testing-library": "^5.4.4",
     "redux-mock-store": "^1.5.3",
     "resolve-url-loader": "^2.3.0",
     "rimraf": "2.6.2",
