+++ src/react-apps/applications/ux-editor/package.json
@@ -28,8 +28,8 @@
     "dot-object": "^1.7.0",
     "immutability-helper": "2.7.0",
     "lodash.isequal": "^4.5.0",
-    "react": "16.6.3",
-    "react-dom": "16.6.3",
+    "react": "16.7.0",
+    "react-dom": "16.7.0",
     "react-grid-layout": "0.16.6",
     "react-hot-loader": "4.3.1",
     "react-modal": "3.4.5",
@@ -51,8 +51,8 @@
     "@types/enzyme": "3.1.15",
     "@types/jest": "^23.3.9",
     "@types/lodash.isequal": "^4.5.3",
-    "@types/react": "16.4.14",
-    "@types/react-dom": "16.0.7",
+    "@types/react": "16.7.18",
+    "@types/react-dom": "16.0.11",
     "@types/react-modal": "3.1.2",
     "@types/react-redux": "6.0.2",
     "@types/react-router-dom": "4.3.1",
