+++ src/react-apps/package.json
@@ -19,8 +19,8 @@
     "jsdom": "^13.1.0",
     "jest-junit": "^5.2.0",
     "lerna": "^3.4.3",
-    "react": "16.6.3",
-    "react-dom": "16.6.3",
+    "react": "16.7.0",
+    "react-dom": "16.7.0",
     "ts-jest": "^23.10.5",
     "typescript": "3.1.6",
     "webpack-cli": "3.0.3"
