+++ src/react-apps/templates/template-app/package.json
@@ -18,7 +18,7 @@
   "license": "3-Clause BSD",
   "dependencies": {
     "babel-core": "6.26.3",
-    "react": "16.6.3",
+    "react": "16.7.0",
     "react-grid-layout": "0.16.6",
     "react-hot-loader": "4.3.1",
     "react-modal": "3.4.5",
@@ -35,7 +35,7 @@
   "devDependencies": {
     "@types/enzyme": "3.1.15",
     "@types/jest": "^23.3.9",
-    "@types/react": "16.4.14",
+    "@types/react": "16.7.18",
     "@types/react-modal": "3.1.2",
     "@types/react-redux": "6.0.2",
     "@types/react-router-dom": "4.3.1",
